GC bug fix
Implemented a callback to the application-side ObjectRegistry for the GC sweep to check before deleting an entity.
Also, storers are kept alive until the next sweep is completed to guarantee at least one strong reference to everything they reference locally / stored.
Tested to work as intended and not cause regression problems.